requirements.txt
# Core dependencies
selenium>=4.0.0
beautifulsoup4>=4.9.3
requests>=2.26.0
langchain>=0.1.0
langchain-openai
python-dotenv>=0.19.0
webdriver-manager>=3.8.0

pytest
pytest-cov
flake8
black
isort
mypy
pre-commit
sphinx
sphinx-rtd-theme
safety
bandit
langchain_openai
openai


# Testing utilities
pytest
pytest-html

# Web parsing
lxml
html5lib

# Utilities
tqdm
pyyaml
openai>=1.0.0
retry>=0.9.2
